```
create 5000 new messages: 1.242s
validate 5000 messages ssb-buttwoo: 517.988ms
validate 5000 messages sbot: 585.584ms
native to db format 5000 messages ssb-buttwoo: 166.615ms
native to db format 5000 messages sbot: 128.726ms
db to native format 5000 messages ssb-buttwoo: 134.514ms
db to native format 5000 messages sbot: 193.874ms
```