@arj03 2022-06-10
This is written a bit quickly. I don't have so much time today. But wanted to put it here as it can make it easier to figure out why EBT is not as fast as before.

db to native format 5000 messages sbot: 193.874ms <- a lot of this overhead is finding the format (around 40-50ms)