@arj03 2022-06-10
The latest commit should better match what is going on when replicating over ebt:

```
create 5000 new messages: 1.204s
validate 5000 messages ssb-buttwoo: 509.465ms
validate 5000 messages sbot: 587.652ms
native to db format 5000 messages ssb-buttwoo: 164.59ms
native to db format 5000 messages sbot: 149.172ms
db to native format 5000 messages ssb-buttwoo: 138.734ms
db to native format 5000 messages sbot: 188.189ms
```