@staltz 2022-06-10
@arj03 These benchmarks were useful, I found out that too much work was done with `SSBURI.decompose` and that basically in many cases we were converting from BFE to URI and back to BFE, so I made quicker versions which just do the thing directly.

Now validation with feed-formats/buttwoo-v1 is slightly faster than validation in ssb-buttwoo:

**Run 1**

```
create 5000 new messages: 1.088s
validate 5000 messages ssb-buttwoo: 522.733ms
validate 5000 messages sbot: 473.532ms
```

**Run 2**

```
create 5000 new messages: 944.712ms
validate 5000 messages ssb-buttwoo: 460.944ms
validate 5000 messages sbot: 463.95ms
```

**Run 3**

```
create 5000 new messages: 997.205ms
validate 5000 messages ssb-buttwoo: 470.361ms
validate 5000 messages sbot: 412.971ms
```

I put the commit https://github.com/ssbc/ssb-db2/pull/351/commits/4c3e0c7133de5b638ed8cff1a2862a75bb56b9c6 here in this PR. We can merge at some point, but thought it would make sense to put all the performance stuff (measurements and optimizations) here first.