@staltz 2022-06-10
Another optimization: https://github.com/ssbc/ssb-db2/pull/351/commits/f9af0005580e9ba8ef64c15af606a2db5f3567b4 I copy-pasted ssb-buttwoo's bipfToButt2 internals, and that made a bit faster.

It's still slower, but there are two things to be done. bipfToButt2 seems slightly incorrect because the `signature` from the database is a buffer, when it should be a string. So my feed-formats/buttwoo-v1 is doing one conversion of string to buffer that bipfToButt2 is not doing. And then the second thing is that `db.findFeedFormatForAuthor(author)` going on, which I think we can get rid of if we pipe the feed format along the function calls.

```
create 5000 new messages: 946.073ms
validate 5000 messages ssb-buttwoo: 411.721ms
validate 5000 messages sbot: 406.933ms
native to db format 5000 messages ssb-buttwoo: 143.907ms
native to db format 5000 messages sbot: 110.885ms
db to native format 5000 messages ssb-buttwoo: 114.169ms
db to native format 5000 messages sbot: 130.072ms
```