@staltz 2022-06-12
Alright, now with commit https://github.com/ssbc/ssb-db2/pull/351/commits/6027e09979c0e15a1f104f4ad3994c324200e830 all measurements are faster in the new implementation.

**Run 1**

```
create 5000 new messages: 933.821ms
validate 5000 messages ssb-buttwoo: 444.749ms
validate 5000 messages sbot: 418.946ms
native to db format 5000 messages ssb-buttwoo: 159.472ms
native to db format 5000 messages sbot: 129.618ms
db to native format 5000 messages ssb-buttwoo: 134.006ms
db to native format 5000 messages sbot: 123.607ms
```

**Run 2**

```
create 5000 new messages: 923.147ms
validate 5000 messages ssb-buttwoo: 442.23ms
validate 5000 messages sbot: 426.279ms
native to db format 5000 messages ssb-buttwoo: 160.077ms
native to db format 5000 messages sbot: 125.565ms
db to native format 5000 messages ssb-buttwoo: 131.653ms
db to native format 5000 messages sbot: 121.745ms
```

**Run 3**

```
create 5000 new messages: 965.35ms
validate 5000 messages ssb-buttwoo: 440.03ms
validate 5000 messages sbot: 413.354ms
native to db format 5000 messages ssb-buttwoo: 161.078ms
native to db format 5000 messages sbot: 127.85ms
db to native format 5000 messages ssb-buttwoo: 128.372ms
db to native format 5000 messages sbot: 118.041ms
```