@arj03 2022-06-12
This is certainly better, in ebt I'm at around 6800. So there is still something there. I'll have a proper look tonight or tomorrow. Think we are close, there has to be something else these benchmarks don't show, but at least there where some good optimizations here.