@staltz 2022-06-12
@arj03 I have another idea. In the log, all the BIPF messages have the same "schema", regardless of feed format. This is because we need to do that because we use BIPF seeking for the indexes. But the `signature` field is never used for BIPF seeking, nor queries nor indexes. The signature field is only relevant for the feed format, during validation. So we could allow the `signature` field of a BIPF-ified buttwoo (i.e. `fromNativeMsg(nativeMsg, 'bipf')`) to be a buffer, and this means we can skip all the "string signature to buffer signature" conversions, and gain in perf.