@staltz 2022-06-12
> This is certainly better, in ebt I'm at around 6800.

Also, (I'm sure you know about this trick, but if we are lucky, you forgot about it) make sure to delete node_modules and reinstall because ssb-db2 is installed via a git branch and gotta make sure the right commit was downloaded. Just in case