@arj03 2022-06-12
> re signature

Yes. ssb-buttwoo stored is as a buffer because of this. We anyway store things like tag and contentHash as a buffer, so if we want a pretty printer for view source / txt export, it has to handle those fields anyway.