@staltz 2022-06-12
@arj03 Can we merge this PR into formats-split? (PS: I added some new measurement utils, `startMeasure`/`endMeasure` to the benchmark script)